Add a basic user admin area.
This will allow us to control admin access. For now the user
admin area is not protected in any way, so that we can let the
board and officers see it without needing to go through the "I signed in
can you give me admin access?" dance for everybody. Before merging
this to master and deploying to produciton we'll add a permission
check for the admin area itself.